Refactor aea to extract protocols and channels.
