+++ sandbox/playground.py
@@ -33,8 +33,8 @@ from typing import Dict, Optional
 import docker
 
 from tac.aea.dialogue.base import Dialogue
-from tac.aea.mail.messages import FIPAMessage
-from tac.aea.mail.protocol import Envelope
+from tac.aea.mail.base import Envelope
+from tac.aea.protocols.fipa.message import FIPAMessage
 from tac.aea.protocols.fipa.serialization import FIPASerializer
 from tac.agents.participant.v1.examples.baseline import BaselineAgent
 from tac.agents.participant.v1.examples.strategy import BaselineStrategy
