+++ tac/aea/mail/base.py
@@ -25,10 +25,126 @@ from abc import abstractmethod
 from queue import Queue
 from typing import Optional
 
-from tac.aea.mail.messages import Address, ProtocolId
-from tac.aea.mail.protocol import Envelope
+from tac.aea.mail import base_pb2
 
 logger = logging.getLogger(__name__)
+
+Address = str
+ProtocolId = str
+
+
+class Envelope:
+    """The top level message class."""
+
+    def __init__(self, to: Optional[Address] = None,
+                 sender: Optional[Address] = None,
+                 protocol_id: Optional[ProtocolId] = None,
+                 message: Optional[bytes] = b""):
+        """
+        Initialize a Message object.
+
+        :param to: the public key of the receiver.
+        :param sender: the public key of the sender.
+        :param protocol_id: the protocol id.
+        :param message: the protocol-specific message
+        """
+        self._to = to
+        self._sender = sender
+        self._protocol_id = protocol_id
+        self._message = message
+        assert type(self._to) == str or self._to is None
+        try:
+            if self._to is not None and type(self._to) == str:
+                self._to.encode('utf-8')
+        except Exception:
+            assert False
+
+    @property
+    def to(self) -> Address:
+        """Get public key of receiver."""
+        return self._to
+
+    @to.setter
+    def to(self, to: Address) -> None:
+        """Set public key of receiver."""
+        self._to = to
+
+    @property
+    def sender(self) -> Address:
+        """Get public key of sender."""
+        return self._sender
+
+    @sender.setter
+    def sender(self, sender: Address) -> None:
+        """Set public key of sender."""
+        self._sender = sender
+
+    @property
+    def protocol_id(self) -> Optional[ProtocolId]:
+        """Get protocol id."""
+        return self._protocol_id
+
+    @protocol_id.setter
+    def protocol_id(self, protocol_id: ProtocolId) -> None:
+        """Set the protocol id."""
+        self._protocol_id = protocol_id
+
+    @property
+    def message(self) -> bytes:
+        """Get the Message."""
+        return self._message
+
+    @message.setter
+    def message(self, message: bytes) -> None:
+        """Set the message."""
+        self._message = message
+
+    def __eq__(self, other):
+        """Compare with another object."""
+        return isinstance(other, Envelope) \
+            and self.to == other.to \
+            and self.sender == other.sender \
+            and self.protocol_id == other.protocol_id \
+            and self._message == other._message
+
+    def encode(self) -> bytes:
+        """
+        Encode the envelope.
+
+        :return: the encoded envelope.
+        """
+        envelope = self
+        envelope_pb = base_pb2.Envelope()
+        if envelope.to is not None:
+            envelope_pb.to = envelope.to
+        if envelope.sender is not None:
+            envelope_pb.sender = envelope.sender
+        if envelope.protocol_id is not None:
+            envelope_pb.protocol_id = envelope.protocol_id
+        if envelope.message is not None:
+            envelope_pb.message = envelope.message
+
+        envelope_bytes = envelope_pb.SerializeToString()
+        return envelope_bytes
+
+    @classmethod
+    def decode(cls, envelope_bytes: bytes) -> 'Envelope':
+        """
+        Decode the envelope.
+
+        :param envelope_bytes: the bytes to be decoded.
+        :return: the decoded envelope.
+        """
+        envelope_pb = base_pb2.Envelope()
+        envelope_pb.ParseFromString(envelope_bytes)
+
+        to = envelope_pb.to if envelope_pb.to else None
+        sender = envelope_pb.sender if envelope_pb.sender else None
+        protocol_id = envelope_pb.protocol_id if envelope_pb.protocol_id else None
+        message = envelope_pb.message if envelope_pb.message else None
+
+        envelope = Envelope(to=to, sender=sender, protocol_id=protocol_id, message=message)
+        return envelope
 
 
 class InBox(object):
