+++ tac/aea/protocols/tac/message.py
@@ -1,0 +1,133 @@
+# -*- coding: utf-8 -*-
+
+# ------------------------------------------------------------------------------
+#
+#   Copyright 2018-2019 Fetch.AI Limited
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ------------------------------------------------------------------------------
+
+"""This module contains the default message definition."""
+from enum import Enum
+from typing import Dict, Optional
+
+from tac.aea.protocols.base.message import Message
+
+
+class TACMessage(Message):
+    """The TAC message class."""
+
+    protocol_id = "tac"
+
+    class Type(Enum):
+        """TAC Message types."""
+
+        REGISTER = "register"
+        UNREGISTER = "unregister"
+        TRANSACTION = "transaction"
+        GET_STATE_UPDATE = "get_state_update"
+        CANCELLED = "cancelled"
+        GAME_DATA = "game_data"
+        TRANSACTION_CONFIRMATION = "transaction_confirmation"
+        STATE_UPDATE = "state_update"
+        TAC_ERROR = "tac_error"
+
+        def __str__(self):
+            """Get string representation."""
+            return self.value
+
+    class ErrorCode(Enum):
+        """This class defines the error codes."""
+
+        GENERIC_ERROR = 0
+        REQUEST_NOT_VALID = 1
+        AGENT_PBK_ALREADY_REGISTERED = 2
+        AGENT_NAME_ALREADY_REGISTERED = 3
+        AGENT_NOT_REGISTERED = 4
+        TRANSACTION_NOT_VALID = 5
+        TRANSACTION_NOT_MATCHING = 6
+        AGENT_NAME_NOT_IN_WHITELIST = 7
+        COMPETITION_NOT_RUNNING = 8
+        DIALOGUE_INCONSISTENT = 9
+
+    _from_ec_to_msg = {
+        ErrorCode.GENERIC_ERROR: "Unexpected error.",
+        ErrorCode.REQUEST_NOT_VALID: "Request not recognized",
+        ErrorCode.AGENT_PBK_ALREADY_REGISTERED: "Agent pbk already registered.",
+        ErrorCode.AGENT_NAME_ALREADY_REGISTERED: "Agent name already registered.",
+        ErrorCode.AGENT_NOT_REGISTERED: "Agent not registered.",
+        ErrorCode.TRANSACTION_NOT_VALID: "Error in checking transaction",
+        ErrorCode.TRANSACTION_NOT_MATCHING: "The transaction request does not match with a previous transaction request with the same id.",
+        ErrorCode.AGENT_NAME_NOT_IN_WHITELIST: "Agent name not in whitelist.",
+        ErrorCode.COMPETITION_NOT_RUNNING: "The competition is not running yet.",
+        ErrorCode.DIALOGUE_INCONSISTENT: "The message is inconsistent with the dialogue."
+    }  # type: Dict[ErrorCode, str]
+
+    def __init__(self, tac_type: Optional[Type] = None,
+                 **kwargs):
+        """
+        Initialize.
+
+        :param tac_type: the type of TAC message.
+        """
+        super().__init__(type=str(tac_type), **kwargs)
+
+    def check_consistency(self) -> bool:
+        """Check that the data is consistent."""
+        try:
+            assert self.is_set("type")
+            tac_type = TACMessage.Type(self.get("type"))
+            if tac_type == TACMessage.Type.REGISTER:
+                assert self.is_set("agent_name")
+            elif tac_type == TACMessage.Type.UNREGISTER:
+                pass
+            elif tac_type == TACMessage.Type.TRANSACTION:
+                assert self.is_set("transaction_id")
+                assert self.is_set("is_sender_buyer")
+                assert self.is_set("counterparty")
+                assert self.is_set("amount")
+                amount = self.get("amount")
+                assert amount >= 0
+                assert self.is_set("quantities_by_good_pbk")
+                quantities_by_good_pbk = self.get("quantities_by_good_pbk")
+                assert len(quantities_by_good_pbk.keys()) == len(set(quantities_by_good_pbk.keys()))
+                assert all(quantity >= 0 for quantity in quantities_by_good_pbk.values())
+            elif tac_type == TACMessage.Type.GET_STATE_UPDATE:
+                pass
+            elif tac_type == TACMessage.Type.CANCELLED:
+                pass
+            elif tac_type == TACMessage.Type.GAME_DATA:
+                assert self.is_set("money")
+                assert self.is_set("endowment")
+                assert self.is_set("utility_params")
+                assert self.is_set("nb_agents")
+                assert self.is_set("nb_goods")
+                assert self.is_set("tx_fee")
+                assert self.is_set("agent_pbk_to_name")
+                assert self.is_set("good_pbk_to_name")
+            elif tac_type == TACMessage.Type.TRANSACTION_CONFIRMATION:
+                assert self.is_set("transaction_id")
+            elif tac_type == TACMessage.Type.STATE_UPDATE:
+                assert self.is_set("initial_state")
+                assert self.is_set("transactions")
+            elif tac_type == TACMessage.Type.TAC_ERROR:
+                assert self.is_set("error_code")
+                error_code = self.get("error_code")
+                assert error_code in set(self.ErrorCode)
+            else:
+                raise ValueError("Type not recognized.")
+        except (AssertionError, ValueError):
+            return False
+
+        return True
