+++ tac/agents/controller/base/reactions.py
@@ -28,8 +28,7 @@ import logging
 
 from tac.aea.agent import Liveness
 from tac.aea.crypto.base import Crypto
-from tac.aea.mail.base import MailBox
-from tac.aea.mail.protocol import Envelope
+from tac.aea.mail.base import MailBox, Envelope
 from tac.aea.protocols.oef.serialization import OEFSerializer
 from tac.agents.controller.base.interfaces import OEFReactionInterface
