+++ tac/agents/participant/v1/base/reactions.py
@@ -32,10 +32,11 @@ from typing import List
 
 from tac.aea.agent import Liveness
 from tac.aea.crypto.base import Crypto
-from tac.aea.mail.base import MailBox
-from tac.aea.mail.messages import FIPAMessage, Address, Message, DefaultMessage
-from tac.aea.mail.protocol import Envelope
+from tac.aea.mail.base import MailBox, Envelope, Address
+from tac.aea.protocols.base.message import Message
+from tac.aea.protocols.default.message import DefaultMessage
 from tac.aea.protocols.default.serialization import DefaultSerializer
+from tac.aea.protocols.fipa.message import FIPAMessage
 from tac.aea.protocols.fipa.serialization import FIPASerializer
 from tac.agents.participant.v1.base.dialogues import Dialogue
 from tac.agents.participant.v1.base.game_instance import GameInstance, GamePhase
