+++ tests/test_messages/test_base.py
@@ -18,8 +18,9 @@
 # ------------------------------------------------------------------------------
 
 """This module contains the tests of the messages module."""
-from tac.aea.mail.messages import Message
-from tac.aea.mail.protocol import ProtobufSerializer, JSONSerializer, Envelope
+from tac.aea.mail.base import Envelope
+from tac.aea.protocols.base.message import Message
+from tac.aea.protocols.base.serialization import ProtobufSerializer, JSONSerializer
 
 
 class TestDefaultSerializations:
