+++ tests/test_messages/test_fipa.py
@@ -20,10 +20,9 @@
 """This module contains the tests for the FIPA protocol."""
 from oef.schema import Description
 
-from tac.aea.mail.messages import FIPAMessage
+from tac.aea.mail.base import Envelope
+from tac.aea.protocols.fipa.message import FIPAMessage
 from tac.aea.protocols.fipa.serialization import FIPASerializer
-
-from tac.aea.mail.protocol import Envelope
 
 
 def test_fipa_cfp_serialization():
